+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # {{next-version}}
 
+* Deprecate `execute()` in favor of `executeAsync()`
+* Deprecate `goSub()` in favor of `executeSync()`
+* Deprecate `return()` in favor of `endAllNodes()`
+
+Deprecated methods will be removed once a plugin with proper aliasing is supplied in another
+package, or when Matthew decides he can live with the new names.
+
 * Plugins can alias user API functions
 * Remove Deprecated `videoPlay` alias of `playVideo`
 * Rename `.goto()` -> `.goToNode()`