+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # {{next-version}}
 
+* Add some universal, base styles that show the experience properly.
 * Deprecate `execute()` in favor of `executeAsync()`
 * Deprecate `goSub()` in favor of `executeSync()`
 * Deprecate `return()` in favor of `endAllNodes()`