+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # {{next-version}}
 
+
+## v0.5.0
+
 * add missing documentation for `wait()`
 * remove documentation for unimplemented timestamps options for `playVideo()`
 * Add some universal, base styles that show the experience properly.